employemanagement/requirements.txt
Django==3.0.5
django-cors-headers==3.2.1
django-storages==1.9.1
djangorestframework==3.11.0

